\name{ayers}
\alias{ayers}
\docType{data}
\title{Birth trauma data}
\description{
  Data from Susan Ayers' (1999) research on trauma after childbirth. These
  are described in Wright and London's (in press) review of multilevel
  modeling. They are used to illustrate different covariance structures.
}
\usage{data(ayers)}
\format{
  A data frame with 287 observations on the following 9 variables.
  \describe{
    \item{\code{partno}}{a numeric vector}
    \item{\code{dep1}}{a numeric vector}
    \item{\code{dep2}}{a numeric vector}
    \item{\code{dep3}}{a numeric vector}
    \item{\code{dep4}}{a numeric vector}
    \item{\code{anx1}}{a numeric vector}
    \item{\code{anx2}}{a numeric vector}
    \item{\code{anx3}}{a numeric vector}
    \item{\code{anx4}}{a numeric vector}
  }
}
\references{
Ayers, S. E. (1999). \emph{Post-traumatic stress disorder following childbirth}.
PhD dissertation. University of London.\cr
 Wright, D.B. & London, K. (in press). Multilevel modelling:
 Beyond the basic applications.
 \emph{British Journal of Mathematical and Statistical Psychology}.
}
\keyword{datasets}
