\name{confcontr}
\alias{confcontr}
\docType{data}
\title{Skagerberg and Wright's (2008) control data}
\description{
  These are the data from the control group in Skagerberg and Wright's
  study of memory conformity. Basically, this is the simplest old/new
  recognition memory design.
}
\usage{data(confcontr)}
\format{
  A data frame with 3100 observations on the following 3 variables.
  \describe{
    \item{\code{subno}}{a numeric vector}
    \item{\code{sayold}}{a numeric vector}
    \item{\code{isold}}{a numeric vector}
  }
}
}
\references{
 Skagerberg, E. M. \& Wright, D. B. (2008). Manipulating power can
 affect memory conformity.
 \emph{Applied Cognitive Psychology}, \bold{22}, 207-216.
}
\examples{
data(confcontr)
}
\keyword{datasets}
