\name{face4}
\alias{face4}
\docType{data}
\title{Horry and Wright's (2008) data}
\description{Horry and Wright's study of the cross race context effect by hits,
false alarms, misses and correct rejections. These can be ised in sdt.}
\usage{data(face4)}
\format{
  A data frame with 45 observations on the following 5 variables.
  \describe{
    \item{\code{subno}}{a numeric vector}
    \item{\code{hits}}{a numeric vector}
    \item{\code{fa}}{a numeric vector}
    \item{\code{misses}}{a numeric vector}
    \item{\code{cr}}{a numeric vector}
  }
}
\references{
 Horry, R. \& Wright, D.B. (2008). I know your face but not where I saw you:
 Context memory is impaired for other race faces.
 \emph{Psychonomic Bulletin & Review}, \bold{15}, 610-614.
}
\examples{
data(face4)
}
\keyword{datasets}
