\name{face4b}
\alias{face4b}
\docType{data}
\title{Horry and Wright (2008) Black faces}
\description{
  Data for the Black faces only from Horry and Wright (2008)
}
\usage{data(face4b)}
\format{
  A data frame with 45 observations on the following 5 variables.
  \describe{
    \item{\code{subno}}{a numeric vector}
    \item{\code{hits}}{a numeric vector}
    \item{\code{fa}}{a numeric vector}
    \item{\code{misses}}{a numeric vector}
    \item{\code{cr}}{a numeric vector}
  }
}
\references{
 Horry, R. & Wright, D.B. (2008). I know your face but not where I saw you:
 Context memory is impaired for other race faces.
 Psychonomic Bulletin & Review, 15, 610-614.
}
\examples{
data(face4b)
}
\keyword{datasets}
