\name{face4w}
\alias{face4w}
\docType{data}
\title{Horry and Wright (2008) - White}
\description{
  Data from Horry and Wright for White faces only
}
\usage{data(face4w)}
\format{
  A data frame with 45 observations on the following 5 variables.
  \describe{
    \item{\code{subno}}{a numeric vector}
    \item{\code{hits}}{a numeric vector}
    \item{\code{fa}}{a numeric vector}
    \item{\code{misses}}{a numeric vector}
    \item{\code{cr}}{a numeric vector}
  }
}
\references{
 Horry, R. & Wright, D.B. (2008). I know your face but not where I saw you:
 Context memory is impaired for other race faces.
 Psychonomic Bulletin & Review, 15, 610-614.
}
\keyword{datasets}
