\name{facec}
\alias{facec}
\docType{data}
\title{Horry and Wright (2008)}
\description{
  Data from Horry and Wright's study of the cross race context effect
}
\usage{data(facec)}
\format{
  A data frame with 7200 observations on the following 4 variables.
  \describe{
    \item{\code{participant}}{a numeric vector}
    \item{\code{facewhite}}{a numeric vector}
    \item{\code{seenbefore}}{a numeric vector}
    \item{\code{sayold}}{a numeric vector}
  }
}
\references{
  Horry, R. & Wright, D.B. (2008). I know your face but not where I saw you:
  Context memory is impaired for other race faces.
  Psychonomic Bulletin & Review, 15, 610-614.
}
\keyword{datasets}
