\name{fne}
\alias{fne}
\docType{data}
\title{Fear of negative evaluation and memory conformity}
\description{
  These data compare fear of negative evaluation with memory conformity
  using a social recognition memory procedure.
}
\usage{data(fne)}
\format{
  A data frame with 4288 observations on the following 6 variables.
  \describe{
    \item{\code{id}}{a numeric vector}
    \item{\code{FNE}}{a numeric vector}
    \item{\code{face}}{a numeric vector}
    \item{\code{sayold}}{a numeric vector}
    \item{\code{isold}}{a numeric vector}
    \item{\code{other}}{a numeric vector}
  }
}
\references{
  Wright, D.B., London, K., & Waechter, M. (under review). Memory conformity
  and social anxiety among a sample of adolescents.
}
\keyword{datasets}
