\name{memrec}
\alias{memrec}
\docType{data}
\title{Memory Recognition Data}
\description{
  Data from the cross race recognition research of Wright, Boyd
  and Tredoux (2003). These are described in Wright and London (in
  press).
}
\usage{data(memrec)}
\format{
  A data frame with 3000 observations on the following 6 variables.
  \describe{
    \item{\code{face}}{a numeric vector}
    \item{\code{facewhite}}{a numeric vector}
    \item{\code{time}}{a numeric vector}
    \item{\code{partno}}{a numeric vector}
    \item{\code{faceold}}{a numeric vector}
    \item{\code{saysold}}{a numeric vector}
  }
}
\references{
	Wright, D. B., Boyd, C. E., \& Tredoux, C. G. (2003). Inter-racial contact and the own
race bias for face recognition in South Africa and England.
\emph {Applied Cognitive Psychology}, \bold {17}, 365-373.
 Wright, D.B. & London, K. (in press). Multilevel modelling:
 Beyond the basic applications.
 \emph{British Journal of Mathematical and Statistical Psychology}.
}
\keyword{datasets}
