\name{confexp}
\alias{confexp}
\docType{data}
\title{Skagerberg and Wright's (2008) Experimental data}
\description{
  Data from the conditions receiving either the low or high power task in
  Skagerberg and Wright's study of memory conformity.
}
\usage{data(confexp)}
\format{
  A data frame with 5200 observations on the following 5 variables.
  \describe{
    \item{\code{task}}{a numeric vector}
    \item{\code{partno}}{a numeric vector}
    \item{\code{other}}{a numeric vector}
    \item{\code{old}}{a numeric vector}
    \item{\code{sayold}}{a numeric vector}
  }
}
\references{
 Skagerberg, E. M. \& Wright, D. B. (2008). Manipulating power can affect
 memory conformity.
 \emph{Applied Cognitive Psychology}, \bold{22}, 207-216.
}
\examples{
data(confexp)
}
\keyword{datasets}
