\name{sdtalt-package}
\alias{sdtalt-package}
\alias{sdtalt}
\docType{package}
\title{
SDT statistics and Alternatives
~~ sdtalt ~~
}
\description{
Allows users, via sdt, to calculate several statistics (including user defined
ones) from signal detection theory for each individual and summary statistics
(mean, trimmed mean, standard deviation, including confidence intervals). The
user may include covariates with the mlmsdt function.
}
\details{
\tabular{ll}{
Package: \tab sdtalt\cr
Type: \tab Package\cr
Version: \tab 1.0-1\cr
Date: \tab 2009-06-01\cr
License: \tab Unlimited)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Daniel B. Wright
Maintainer: Daniel B. Wright <dwright@fiu.edu>
}
\references{
Wright, D.B., Horry, R., & Skagerberg, E.M. (2009). Functions for
traditional and multilevel approaches to signal detection theory.
Behavior Research Methods, 41, 257-267.
}
\keyword{ package }
